Create classifications from report status
If a report has been sent for review then it is implicitly constructive
(this is the classification it is imported with).

If the person who sent it for review is also a reviewer, then create a
constructive classification for the review.

This will prevent the edit from showing up in the queue for the user and
require 1 (matching) review to classify the edit.

Essentially this is just an effort to reduce the overhead of reviewing
for the backlog.