Create classifications from Huggle
ClueBot currently respects the whitelist from Huggle, if an edit was
done by a Huggle "trusted user" then add a constructive classification.

Note: If this happens to a reported edit, then in combination of
`add_reviews_from_report` there could be no human review in this
interface (1 human review via report interface).

As the bot loads the trusted list and excludes reverts from said list,
there should not be an overlap of reported and trusted, outside of the
case an edit is reverted/reported and then the user becomes trusted
last.