Add command to re-generate training data
Previously there was a bug in `get_user_warning_count` where the page
name was not escaped properly, which resulted in an implicit `0`
warnings.

As we guard the count by edit time (so the view is consistent to what
the bot would see in real time), we need to do this per edit not just
per user.

Note: We exclude deleted edits; for historical (edit set) this will be
correct, but there may be some edits that were since added but have been
deleted... this should be a very small number, so hopefully won't cause
too many issues.