EditSetParser - handle <text></text> as "" not None
If the edit was blanking of the page, then there will be no text.

We have a check in WPRevision to assert `text` is not None, i.e. it has
been set on the model.

Making `text` "" will ensure `has_complete_training_data` returns True
and we store the edit data.